+++ HISTORY.rst
@@ -9,6 +9,7 @@ development
 
 - improvements to docs
 - added slack API example
+- ``related`` decorated replaces ``Relation`` query class
 
 0.5.0 (2018-01-30)
 ++++++++++++++++++